feat: added a way to view bookings to your venues through detailedView if you are the owner. restyled header and optimized profile code a bit
