+++ src/components/Layout/Header.jsx
@@ -2,6 +2,7 @@ import { useContext, useEffect, useState, useRef } from "react";
 import { useNavigate, useSearchParams, Link } from "react-router-dom";
 import { UserContext } from "../../utility/UserContext";
 import logo from "../../assets/images/stay-and-trip.png";
+import { ArrowUp, ArrowDown } from "lucide-react";
 
 export function Header() {
     const { user, setUser } = useContext(UserContext);
@@ -18,6 +19,18 @@ export function Header() {
         setSearchTerm(searchParams.get("q") || "");
         setSortOrder(searchParams.get("sortOrder") || "asc");
     }, [searchParams]);
+
+    useEffect(() => {
+        const handleClickOutside = (e) => {
+            if (menuRef.current && !menuRef.current.contains(e.target)) {
+                setMenuOpen(false);
+            }
+        };
+        document.addEventListener("mousedown", handleClickOutside);
+        return () => {
+            document.removeEventListener("mousedown", handleClickOutside);
+        };
+    }, []);
 
     const handleSearch = (e) => {
         e.preventDefault();
@@ -48,7 +61,11 @@ export function Header() {
         >
             <div className="flex items-center justify-between max-w-[1600px] mx-auto p-4">
                 <div className="flex-shrink-0">
-                    <Link to="/" aria-label="Home" className="h-16">
+                    <Link
+                        to="/"
+                        aria-label="Home"
+                        className="flex items-center h-16"
+                    >
                         <img
                             src={logo}
                             alt="Stay and trip Logo"
@@ -58,13 +75,13 @@ export function Header() {
                 </div>
                 <form
                     onSubmit={handleSearch}
-                    className="flex items-stretch flex-1 mx-4 max-w-lg"
+                    className="flex items-stretch flex-1 mx-4 max-w-lg shadow-sm rounded-md overflow-hidden"
                     role="search"
                     aria-label="Site search"
                 >
                     <input
                         type="text"
-                        className="flex-grow px-4 py-2 border-t border-b border-l border-gray-300 rounded-l-md focus:outline-none"
+                        className="flex-grow px-4 py-2 border-0 focus:outline-none"
                         placeholder="Search title or description..."
                         value={searchTerm}
                         onChange={(e) => setSearchTerm(e.target.value)}
@@ -72,30 +89,35 @@ export function Header() {
                     />
                     <button
                         type="submit"
-                        className="px-4 py-2 bg-blue-500 text-white rounded-r-md hover:bg-blue-600 cursor-pointer"
+                        className="px-4 py-2 bg-blue-500 text-white hover:bg-blue-600 cursor-pointer"
                         aria-label="Submit search"
                     >
                         Search
                     </button>
                 </form>
-
                 <button
                     onClick={toggleSort}
-                    className="mr-4 px-4 py-2 border border-gray-300 rounded hover:bg-gray-100 cursor-pointer"
+                    className="mr-4 px-4 py-2 bg-gray-100 text-gray-700 rounded hover:bg-gray-200 flex items-center gap-1"
                     aria-label={
                         sortOrder === "asc"
                             ? "Sort descending"
                             : "Sort ascending"
                     }
                 >
-                    {sortOrder === "asc" ? "Asc ↑" : "Desc ↓"}
+                    {sortOrder === "asc" ? (
+                        <ArrowUp size={16} />
+                    ) : (
+                        <ArrowDown size={16} />
+                    )}
+                    {sortOrder === "asc" ? "Asc" : "Desc"}
                 </button>
-
                 {user ? (
                     <div className="relative" ref={menuRef}>
                         <button
                             onClick={() => setMenuOpen((o) => !o)}
-                            className="p-0 border-0 bg-transparent"
+                            className={`p-0 border-0 bg-transparent hover:opacity-80 transition-opacity ${
+                                menuOpen ? "ring-2 ring-blue-500" : ""
+                            }`}
                             aria-haspopup="menu"
                             aria-expanded={menuOpen}
                             aria-label="User menu"
@@ -107,56 +129,57 @@ export function Header() {
                                 className="h-8 w-8 rounded-full cursor-pointer"
                             />
                         </button>
-                        {menuOpen && (
-                            <div
-                                id="user-menu"
-                                className="absolute right-0 mt-2 w-48 bg-white border border-gray-200 shadow-lg rounded"
-                                role="menu"
-                                aria-label="User options"
-                            >
-                                <ul>
+                        <div
+                            id="user-menu"
+                            className={`absolute right-0 mt-2 w-48 bg-white border border-gray-200 shadow-lg rounded transition-opacity duration-200 ${
+                                menuOpen
+                                    ? "opacity-100"
+                                    : "opacity-0 pointer-events-none"
+                            }`}
+                            role="menu"
+                            aria-label="User options"
+                        >
+                            <ul>
+                                <li>
+                                    <Link
+                                        to={`/profile/${user.name}`}
+                                        className="block px-4 py-2 hover:bg-gray-100"
+                                        role="menuitem"
+                                    >
+                                        My Profile
+                                    </Link>
+                                </li>
+                                {venueManager && (
                                     <li>
                                         <Link
-                                            to={`/profile/${user.name}`}
+                                            to="/venues/createVenue"
                                             className="block px-4 py-2 hover:bg-gray-100"
                                             role="menuitem"
                                         >
-                                            My Profile
+                                            Create Listing
                                         </Link>
                                     </li>
-                                    {venueManager && (
-                                        <li>
-                                            <Link
-                                                to="/venues/createVenue"
-                                                className="block px-4 py-2 hover:bg-gray-100"
-                                                role="menuitem"
-                                            >
-                                                Create Listing
-                                            </Link>
-                                        </li>
-                                    )}
-
-                                    <li>
-                                        <button
-                                            onClick={handleLogout}
-                                            className="w-full text-left px-4 py-2 hover:bg-gray-100 cursor-pointer"
-                                            role="menuitem"
-                                        >
-                                            Logout
-                                        </button>
-                                    </li>
-                                </ul>
-                            </div>
-                        )}
+                                )}
+                                <li>
+                                    <button
+                                        onClick={handleLogout}
+                                        className="w-full text-left px-4 py-2 hover:bg-gray-100 cursor-pointer"
+                                        role="menuitem"
+                                    >
+                                        Logout
+                                    </button>
+                                </li>
+                            </ul>
+                        </div>
                     </div>
                 ) : (
                     <div className="flex-shrink-0">
                         <Link
                             to="/auth"
-                            className="mr-4 px-4 py-2 text-blue-500 hover:underline"
-                            aria-label="Login"
+                            className="mr-4 px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
+                            aria-label="Login or Register"
                         >
-                            Login/Register
+                            Login / Register
                         </Link>
                     </div>
                 )}
